fix: sqlx pre-commit-hook
The hook was failing when running `nix flake check` due to the flake
sandbox. It turns out that setting `enable = false` on the hook causes
pre-commit-hooks.nix to skip that check during sandbox builds. This
means that the hook is able to catch issues in code before it is
committed, but does not cause issues for CI.